trying out to see if viewmodels as modules make sense
my original motivation for this is to put view models below the
fragment/views in the component hierarchy so that a view model could
use fragment/view-dependent objects without having them as constructor-
injected final members. it only works if there are no actions that
span multiple fragment lives. meaning no concurrency. i.e., it's
pretty much useless. it does have an advantage that you don't need
to make the view models dagger-reachable so you can use
SavedStateHandle for example. other than that, there's not much
advantage and it comes with a bit more boilerplate.